pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Sergio Lema <serlesen@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Flask = "^2.0.2"
Flask-HTTPAuth = "^4.5.0"
PyJWT = "^2.3.0"
Flask-SQLAlchemy = "^2.5.1"
psycopg2 = "^2.9.3"
marshmallow-sqlalchemy = "^0.27.0"
pytest = "^7.0.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
